refactor: split accessibility guideline files by responsibility
- Move ViolationNode and GuidelineViolation into their own files; guideline.dart
  now only holds the WidgetbookGuideline interface and evaluateGuidelines.
- Split native guidelines into one file per guideline under native/, with shared
  geometry/format helpers in native/guideline_utils.dart and a native barrel.
- Add an accessibility.dart barrel so widgetbook.dart exports a single entry.
- Drop the now-redundant guidelines re-export from test.dart.
- Pass scenario violations as serialized maps (List<Map<String, dynamic>>),
  matching how semanticsData is handed to ScenarioMetadata.
